Add parallel input/output support for "GPIO Manual Control"
Initial commit to expand the GPIO app's "manual control" functionality
and make it actually useful for hardware tinkering.  Show the state of
all available pins at once and allow them to be configured individually
as inputs or outputs.  Allow outputs to be used as momentary buttons
with "OK" or as toggle switches with "up"/"down".